refactor: move the command bar's wiring out of the layout
`vmux_layout/src/command_bar/` was 3,600 lines of launcher inside the window
shell. The layout owned it because the launcher needs to know what tabs, stacks
and panes exist - but owning the whole module to answer that question is the
wrong trade, and it put the layout in the position of implementing a surface it
only hosts.

It could not simply move to `vmux_command`: the module reads the layout's model,
so that would invert the dependency into a cycle. It belongs in `vmux_browser`,
whose job is composing pages into the shell, and which already depends on both.

Splitting it turned up a third thing. `build_command_bar_open_payload` is
documented as "shared by the command-bar modal and the home launcher", and that
is exactly true: `vmux_layout::start` builds command bar payloads too. So the
payload assembly is not the modal's, and not the layout's - it is the launcher's,
and every input it takes was already a `vmux_command`, `vmux_wire` or `vmux_ui`
type. The only workspace-shaped input is a `Vec<CommandBarTab>`.

So it lands in three places, each holding what only it can answer:

  vmux_command/host/payload.rs      the payload both surfaces send
  vmux_layout/host/workspace_snapshot.rs   what tabs and stacks exist
  vmux_browser/command_bar/         the wiring between them

`vmux_layout` no longer mentions the command bar, and no cycle is introduced:
both consumers sit downstream of `vmux_command`, and `vmux_browser` sits
downstream of both.
