+++ crates/page/vmux_command/src/host.rs
@@ -27,6 +27,7 @@ pub mod command;
 pub mod issued;
 pub mod open;
 pub mod page_key;
+pub mod payload;
 pub mod shortcut;
 pub mod snapshot;
 
@@ -35,4 +36,8 @@ pub use command::*;
 pub use issued::{CommandIssued, CommandIssuer};
 pub use open::*;
 pub use page_key::{PageKeyPlugin, ScopedKeys};
+pub use payload::{
+    CommandBarEntry, build_command_bar_open_payload, command_bar_open_payload, command_list,
+    localized_command_name,
+};
 pub use snapshot::*;
