--- crates/page/vmux_layout/src/command_bar.rs
@@ -1,10 +1,0 @@
-#[cfg(host)]
-pub mod handler;
-#[cfg(host)]
-pub mod key;
-#[cfg(host)]
-pub mod panel;
-#[cfg(host)]
-pub mod state;
-#[cfg(host)]
-pub mod work_snapshot;
