+++ crates/page/vmux_layout/src/lib.rs
@@ -6,7 +6,6 @@
     clippy::new_ret_no_self
 )]
 
-pub mod command_bar;
 pub mod event;
 pub mod protocol;
 pub mod reconcile;
