+++ crates/vmux_browser/Cargo.toml
@@ -22,6 +22,7 @@ crossbeam-channel = { workspace = true }
 image = { version = "0.25", default-features = false, features = ["png"] }
 libc = "0.2"
 polling = { workspace = true }
+rkyv = { workspace = true }
 reqwest = { version = "0.12", default-features = false, features = ["blocking", "rustls-tls"] }
 ron = { workspace = true }
 serde = { workspace = true }
@@ -41,6 +42,7 @@ vmux_setting = { path = "../page/vmux_setting" }
 vmux_space = { path = "../page/vmux_space" }
 vmux_terminal = { path = "../page/vmux_terminal" }
 vmux_ui = { path = "../vmux_ui", default-features = false }
+vmux_wire = { path = "../vmux_wire" }
 
 [target.'cfg(target_os = "macos")'.dependencies]
 objc2-app-kit = { version = "0.3", default-features = false, features = [
