+++ crates/vmux_browser/src/command_bar/state.rs
