+++ crates/vmux_browser/src/command_bar/work_snapshot.rs
