+++ crates/vmux_browser/src/frame_rate.rs
@@ -4,6 +4,7 @@
 //! the hover state and the frame rate describe the frame that just happened rather than the one
 //! before it. The webview idles slowly and bursts back to full rate when the host emits to it.
 
+use crate::command_bar::state::CommandBarStateQuery;
 use bevy::{
     input::{
         keyboard::KeyboardInput,
@@ -18,7 +19,6 @@ use bevy_cef::prelude::*;
 use std::sync::atomic::Ordering;
 use vmux_command::CommandBar;
 use vmux_layout::Browser;
-use vmux_layout::command_bar::state::CommandBarStateQuery;
 use vmux_layout::{
     Header, LayoutCef,
     event::{LAYOUT_COMMAND_BAR_OPEN_EVENT, STACKS_EVENT, TABS_EVENT},
@@ -188,7 +188,7 @@ fn refresh_active_windowed_hover(
     >,
     mut state: Local<WindowedHoverRefreshState>,
 ) {
-    if vmux_layout::command_bar::handler::is_command_bar_open(&modal_q) {
+    if crate::command_bar::handler::is_command_bar_open(&modal_q) {
         *state = WindowedHoverRefreshState::default();
         return;
     }
